requirements.txt
selenium
pyyaml
requests
browsermob-proxy
psutil
brotli